backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
Pillow>=10.0.0
aiofiles>=23.2.0
pydantic[email]>=2.0.0
python-multipart>=0.0.6
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
cryptography>=41.0.0
aiosmtplib>=3.0.0
python-dotenv>=1.0.0
slowapi>=0.1.8
psycopg2-binary>=2.9.0
email-validator>=2.0.0